requirements.txt
streamlit>=1.41.1,<2.0.0
plotly>=6.0.0,<7.0.0
tortoise-orm>=0.24.0,<0.25.0
asyncpg>=0.30.0,<0.31.0
aerich>=0.8.1,<0.9.0
tomlkit>=0.13.2,<0.14.0
langchain>=0.3.17,<0.4.0
smolagents[litellm]>=1.6.0,<2.0.0
langchain-community>=0.3.16,<0.4.0
rank-bm25>=0.2.2,<0.3.0
langchain-openai>=0.3.3,<0.4.0
faiss-cpu>=1.10.0,<2.0.0
pypdf2>=3.0.1,<4.0.0
python-pptx>=1.0.2,<2.0.0
aiosqlite>=0.16.0,<0.21.0
python-dotenv>=1.0.0,<2.0.0 